clayer:  Overriding ToString() of cLayer, instead of using a LayerString() method, as it is cleaner.
cPuzzleBuilder:  Adding logic (currently unfinished) to actually assign sudoku-compliant values to a 2d Sudoku puzzle.  I'm hoping that my algorithms will be expandable to the 3d case.

cSquare:  Adding TrySetValue(), which attempts to randomply assign a value to the square, and ensure that the new value doesn't violate sudoku rules.

g.cs:  Removed the "getter" form the Seed property, making it read-only, as I think this will help clarify it's intent.

Program.cs:  Refactored BuildPuzzle() to BuildPuzzle(int) so that a seed value needs to be passed to the function call.  Also accepting seed values from the command line as an argument.