cSquare: refactored logic in TrySetValue(), by ensuring that we had a value (available) prior to stepping into the Sudoku Rule Validing code, the code is much simpler to follow.
cValidation:
Not unit test code, as the name might imply, but rather a sandbox in which I can stomp around and write code to help me visualize the workings of the classes.  For example, to visualize which squares are indexed by cLayer[i][j].

I added a total of the values "Banner Report" in the "CheckLinks()" routine, which reveals each squares's Validation List Members, by giving them a value and then printing out the layer to the console. (they member's all have non-zero, sequential values, and the non-members are all zeros.

Program:
Deleted some dead code, and am calling cValidate.CheckLinks() in the CreateValidationLinks() routine.

UnitTest1:
Added two TestMethods.
TestLinkSums:
Loops through all of the Squares in each Square's Validation List, giving them
a value starting at 1, and each one incrementing each value by 1.
After assigning the values, it re-iterates through the linked list and verifies that
the sum of all of them = 231 SUM(1...21), as the list length is 21.

TestLinkLengths()
Verifies that the lists are all have 21 members.
TestLinkSums